[Refactor] 코드 에디터 타입 및 상수 정리
